Moved legend logic to service to maintain active layer state
State was being reset when the logic lived in the component, instead of the service.
